refactor: put sampling sites in their own file
